refactor attachMethods function
