+++ src/components/dataTableStatic/utils/index.ts
@@ -1,0 +1,2 @@
+export * from "./attachCellComponentsToMetaData";
+export * from "./attachMethodsToMetaData";
