--- src/components/utils/types.ts
@@ -1,6 +1,0 @@
-import { TopLevelCondition } from "json-rules-engine";
-export interface CustomRuleType {
-  conditions: TopLevelCondition;
-  success: any;
-  failure: any;
-}
